Add explicit test for DirectiveFactory
